requirements.txt
langchain
langgraph
python-dotenv
openai
feedparser
requests
langchain-community
langchain-openai